iot-subscriber/requirements.txt
psycopg2-binary
flask
paho-mqtt
datetime
threaded